Adds detail activity linking

+++ app/src/main/java/uk/dan_gilbert/paytouch/data/model/Actor.java
@@ -12,7 +12,7 @@ public class Actor {
 
     public String name;
     public boolean adult;
-    public long identifier;
+    public int identifier;
     public Film[] knownFor;
     public float popularity;
     public String profilePath;
